Revert "add Baloo DBus signals for moved or removed files"
Summary:
Revert "add missing license"
This reverts commit 96fb5896c304b7292fc3cea87fb584b14b270448.

This reverts commit ef0e268c657718d4a353638f91d3723624c358ac.

The BalooWatcherApplication DBus interface was created to be used by
Elisa, but the only use there is to create some debug output.
As soon as Elisa registers itself, each rename causes a bunch
of DB queries (in case a directory is renamed, it builds the list
of all children) and a (potentially huge) DBus call.

The DBus interface is totally awkward, as it calls a method in the
watcher application (i.e. replicating the message on the sender side)
instead of using DBus signals and letting the DBus daemon do its job.

Keep the DBus interface as a stub, but remove the caused overhead.

Test Plan: ctest

Reviewers: #baloo, mgallien, ngraham

Reviewed By: #baloo, mgallien, ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks, #baloo

Differential Revision: https://phabricator.kde.org/D28980