[Engine] Propagate transaction errors
In case a transaction can not be commited, exit the extractor process. If
the commit only failed because it is too large the extractor is started
again with half the batch.

If the batch only contains a single file and marking it as failed is not
possible we are stuck, so also exit from the main process.

CCBUG: 425017