[AdvancedQueryParser] Fix out-of-bound access if last character is a comparator
If the very last character of the search string is one of '<', '>', ':'
or '=', the index for QString::at is to large by 1, and in Debug builds
a Q_ASSERT will trigger: "uint(i) < uint(size())".

As the index points to the trailing `\0` this OOB access can not be
detected by e.g. Valgrind.