[MetaDataMover] Extend tests to cover renames
The Inotify processing happens asynchronously after the actual filesystem
changes. If a second rename or a delete happens immediately after the
first rename, processing of the first rename has to work without the
state visible without the filesystem.

The `testMoveDeleteFile()` fails as MetaDataMover can not resolve the
file in the filesystem for the move and then fails to delete it from
the index.

The `testMoveTwiceCombined()` only succeeds by chance, as it happily
renames a file it does not know.

Add a corresponding test for KInotify to assert the preconditions in
MetaDataMoverTest match.