[FileIndexerConfig] Fix hidden check for explictly included directories
An explicitly included hidden directory was correcly indexed, but any
directories below were skipped as the full path was checked if some
of it was a hidden directory.

On UNIX-like systems, the check can be skipped as the "hidden" check is
already done in `shouldFileBeIndexed`. On other systems (Windows),
`QFileInfo::isHidden()` is used for all directories below the included
directory.