Do not crash when a dir is move/delete is detected by kinotify
It appears that sometimes with kinotify sends a directory move or delete event,
it may be sent with a trailing slash on the path - and this causes filewatch
to fire an assertion.

Instead we might expect that things like that happen and just avoid tacking
on the required trailing slash ourselves.

BUG: 435143