Revert "Port away from deprecated QDateTime(QDate) constructor"
This causes a failure to build from source on MSVC as pointed out by the CI system.
Because Baloo is a Framework, leaving this broken is not an option as it breaks the ability to complete Dependency Builds and thus the ability to provide Windows CI to all other KDE projects.
Therefore revert this for now until a proper fix can be implemented - preferrably by going through review.

This reverts commit 23c8e99e1c8144472e2f3a1cdd3d9a2603af4193.

CCMAIL: kde-frameworks-devel@kde.org