Properly shutdown extractor process
Depending on the exact timing, the extractor process will either exit due
to a closed pipe, or it will be killed from the parent from the
destructor. Give the chance a child to terminate on its own.

This is more a less a cosmetic issue, as the child exit status is no
longer relevant when the destructor is called, but it removes some
ambiguity when debugging/tracing.