aggressively resource constrain baloo_file
baloo_file is a very particular background service in that it has the
potential to eat all the user's CPU and IO time. To prevent this from
practically ever happening we'll lower the respective weight controls
way down and further constrain memory consumption to make it incredibly
unlikely that we baloo may be the case of OOM scenarios.

this means baloo still gets as much CPU/IO time as it wants, but not
ever at the cost of another cgroup.

I've been running these constraints for well over a year and noticed no
negative effects.