Revert "Also make autostart desktop file versioned"
The service is a unique DBus instance, just renaming it does not change
that, and introduces random behavior in case both versions are
autostarted.

Distributions should install either the KF5 or KF6 version (both are fully
compatible). The service consists of the service file, the actual daemon
(baloo_file) and the extractor helper (baloo_file_extractor).

This reverts commit 47ee218f09e967a36f7b501d534fc8c5bb5c2756.