[DocumentUrlDB] Remove recursive removal of parent directories
In case a directory is only part of the DB for structural integrity
(i.e. unindexed parent directories of an indexed directory), removal
of the indexed directory would also remove the parent directory.

This is fairly unnecessary as the entries do not pose any problem,
they would be removed when the actual directory (or any of its parents)
are removed.

Removal of the up-traversal gets rid of some quite convoluted code.

(cherry picked from commit 10551a53712a029151eca0d0508e32565887d322)