[QueryTest] Add test case for terms folded to empty strings
The termgenerator removes leading, trailing and isolated punctuation
characters. While this is no problem for leading or trailing punctuation
as the normalized term is still non-empty, isolated punctuation ends up
as an empty string.

The matcher then tries to find e.g. "(content:'' OR filename:'')", which
returns not results, instead of the exptedted "everything". Unfortunately
it is non-trivial to remove such Terms.

For the prase test, this is not at problem, as it always sees the whole
phrase and can just skip the problematic part.

CCBUG:407664
(cherry picked from commit c8d1b7cabcc810867f400e5ccfb96e41b03e0635)