Revert "Implement immutable IndexerConfigData"
This reverts commit c2a74b8bc13306734d9d869da0416d388b352ed6.

The change was merged without maintainer approval, and without prior
notice.

The mix of inheritance and aggregation of data used in the
FileIndexerConfig and FileIndexerConfigData, and likewise in the
IndexerConfig/IndexerConfigData makes the whole implementation quite
complex and even harder to understand and maintain.

The current IndexerConfig currently implments two different sets
of function, which should be separate:

1. Interfacing with the actual config
2. Evaluating indexing decisions based on the config

It is even quite questionable if the second set even should be exposed
to the outside, as it is actually information only applicable internally,
and outside users should not query if a document *should* be indexed, but
if the document *has* been indexed. The latter is exposed via
`Baloo::File`.