[KInotify] Removed unused and obsolete `available()` method
The method is never called, so it is fairly pointless, and it is very
unlikely this code is ever called an a kernel version <= 2.6.13.

Also, if we have a valid file descriptor, the kernel apparently
supports inotify.

(cherry picked from commit daf2c211a7ededb75d5571d093c6d6800ac0efdf)