[KInotify] Fix _k_addWatches helper when hitting descriptor limit
Break the loop in case the watch descriptors are exhausted without
emitting the `installedWatches` signal. Previously, it was emitted when
the path list was emptied, but omitted when the loop was left due to
hitting the batch size limit.

The return value is pointless, as the helper is called asynchronously.