[Extractor] Change to QCoreApplication
baloo_file_extractor is not a GUI tool. QGuiApplication may have been a
requirement of file extractors in the past, but it does not seem to be
the case now. All tests in KFileMetadata use Q_GUILESS_MAIN which is a
QCoreApplication.

This will help with some issues if the baloo_file outlives the platform,
and also help towards future sandboxing work.