[Extractor] Explicitly abort the transaction on early exit
In case the extractor quits because its parent exited (signaled by the
the socket event on the pipe), the transaction should be aborted.

Replacing the raw pointer with a unique_ptr would be sufficient to
avoid the leak, but the destructor issues a warning when a write
transaction is not commited/aborted before deleting it, so add an
explicit abort().