[Extractor] Add hint that the file metadata extractor started correctly
baloo_file launches baloo_file_extractor, if the file_extractor crashes
it will get respawned to identify the broken file in a bisect, before
continuuing.

If there is an error, such as the gui platform being down, or a distro
upgrade left baloo-file-metadata-extractor in a broken state the
extractor would be launched an insane amount of times with the same
result.

This patch adds a flag that the extractor started correctly and
intialised the app and platform. If the extractor crashes before this
flag is set, baloo_file bails out to be restarted by the relevant
mechanism which has restart protection.