[excludeMimeTypes] Exclude model/obj and text/rust from content indexing
3D models, mimetype model/obj, are subtypes of plain/text and are content
indexed by default. These files can be large and full of numbers.

Rust source seems be to content indexed by default, by comparison
C and C++ source is excluded on the basis of their mime types.

Excluding these by mimetype means that Baloo will continue to index
the filenames but not the content

(The model/obj mimetype seems to be a recent addtion to the
freedesktop.org.xml list, it may not be present in LTS distros)

CCBUG: 488533