[Database] Propagate failure reasons for open, remove unused functions
This allows to get rid of the isOpen() and isAvailable() auxiliary
functions (only used by the TagListJob). Expecially the "isAvailable()"
function is problematic, as it is prone to TOCTU races, and may also
return incorrect results, i.e. when the file exists, but the DB is
corrupted.

The path() method is also unused, remove it.