[FileContentIndexer] Avoid race on updatedFiles modification
The filecontentindexer is living in the main thread, and thus the
`FileContentIndexer::slotFinishedIndexingFile` is process as a queued
signal, i.e. in the main thread. The filecontentindexer itself is run
in a dedicated thread.

Modifying m_updatedFiles from both threads is racy and incorrect, and
apparently causes crashes.

Replace the m_updatedFiles member with a local variable, and make sure it
is only ever updated by the runner thread, using a Qt::DirectConnection.

Although it is probably safe to modify the m_finishedCount, better move
the complete remaining files/time logic to the TimeEstimator.

SENTRY: BALOO-1C1