Fix casing of ecm_generate_headers PREFIX arg value
The PREFIX expects the actual, usually CamelCase variant of the name.
The lowercase version worked by chance, as the argument is not used
in the actually generated files, and the build internally also does not
use the CamelCase forwarding include directives, for which a subdir by
the name is created in the build dir.