[FileWatchTest][KInotifyTest] Cover atomic file replacing
A file can either be replaced by first deleting it an creating a new
file in its place (or by moving another file), or doing it atomically.

The latter is done using rename/renameat syscall, which is exercised
by QSaveFile.

In case of an atomic file replacement, inotify does not generate
any event for an eventual old file, which is therefor generated by
the metadatamover.

Verify both the inotify behavior as documented above, and the resulting
behavior of the FileWatch.